requirements.txt
arize-phoenix>=0.15.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp>=1.24.0
opentelemetry-instrumentation-logging>=0.45b0
crewai>=0.28.0
langchain-openai>=0.2.0
google-genai>=0.1.0
python-dotenv>=1.0.1
